[feat] #6 domain model 분리
