+++ domain/memo/src/main/java/com/memozi/memo/model/MemoItem.kt
@@ -1,4 +1,4 @@
-package com.memozi.memo
+package com.memozi.memo.model
 
 data class MemoItem(val title: String, val content: String, val date: String)
 
@@ -12,26 +12,5 @@ fun dummyMemoItems(): List<MemoItem> {
         MemoItem("Memo Title 2", "This is the content of memo 2", "2024-08-31"),
         MemoItem("Memo Title 2", "This is the content of memo 2", "2024-08-31"),
         MemoItem("Memo Title 3", "This is the content of memo 3", "2024-08-31")
-    )
-}
-
-data class CategoryItem(
-    val imageUrl: String,
-    val name: String,
-    val textColor: String
-)
-
-fun dummyMemoCategoriesItems(): List<CategoryItem> {
-    return listOf(
-        CategoryItem(
-            imageUrl = "https://github.com/user-attachments/assets/2473834b-2be4-4584-b143-7e7269a6607c",
-            name = "카테고리 1",
-            textColor = "#FFFFFF"
-        ),
-        CategoryItem(
-            imageUrl = "https://github.com/user-attachments/assets/6cf10fec-27cb-4367-ba97-40c937fbb92c",
-            name = "카테고리 2",
-            textColor = "#FF0000"
-        )
     )
 }
